+++ portal/src/pages/accessibility.mdx
@@ -1,0 +1,7 @@
+---
+title: Tilgjengelighet
+---
+
+import NavPage from "../components/NavPage";
+
+<NavPage {...props} />
