+++ portal/src/pages/components.mdx
@@ -1,0 +1,7 @@
+---
+title: Komponenter
+---
+
+import NavPage from "../components/NavPage";
+
+<NavPage {...props} />
