--- portal/src/pages/components.tsx
@@ -1,14 +1,0 @@
-import React, { Fragment } from "react";
-import { Link } from "gatsby";
-
-import SEO from "../components/seo/seo";
-
-const ComponentsPage = () => (
-  <Fragment>
-    <h1>Hi from the second page</h1>
-    <p>Welcome to page 2</p>
-    <Link to="/">Go back to the homepage</Link>
-  </Fragment>
-);
-
-export default ComponentsPage;
