+++ portal/src/pages/patterns.mdx
@@ -1,0 +1,7 @@
+---
+title: Mønster
+---
+
+import NavPage from "../components/NavPage";
+
+<NavPage {...props} />
