--- portal/src/pages/patterns.tsx
@@ -1,15 +1,0 @@
-import React, { Fragment } from "react";
-import { Link } from "gatsby";
-
-import SEO from "../components/seo/seo";
-
-const PatternPage = () => (
-  <Fragment>
-    <SEO title="Page three" />
-    <h1>Hi from the second page</h1>
-    <p>Welcome to page 2</p>
-    <Link to="/">Go back to the homepage</Link>
-  </Fragment>
-);
-
-export default PatternPage;
