+++ portal/src/pages/resources.mdx
@@ -1,0 +1,7 @@
+---
+title: Ressurser
+---
+
+import NavPage from "../components/NavPage";
+
+<NavPage {...props} />
